Move axios interceptor wrapper to component folder
